// src/views/app/reports/components/charts/acquisitionGeneralData/AcquisitionGeneralData.types.ts
export interface AcquisitionGeneralDataPropTypes {
  application: {
    total: number | undefined
    percentage: number | undefined
  }
  selected: {
    total: number | undefined
    percentage: number | undefined
  }
  rejected: {
    total: number | undefined
    percentage: number | undefined
  }
  pending: {
    total: number | undefined
    percentage: number | undefined
  }
  hired: {
    total: number | undefined
    percentage: number | undefined
  }
}